Initial spike to build out challenge API/data
This is pretty basic but it gets the job done for now!

The data could use a bit of work - in particular, it’s very regular and it doesn’t currently have any difference between the broker and carrier slices. Could also use a once-over to confirm that the documentation is reasonable.